# aspnetcore-server-generated/src/IO.Swagger/Controllers/FaultHandlingApi.cs
/*
 * Service-Oriented Vehicle Diagnostic API
 *
 * Definition of ASAM SOVD API
 *
 * OpenAPI spec version: 1.0.0
 * 
 * Generated by: https://github.com/swagger-api/swagger-codegen.git
 */

using System;
using System.Collections.Generic;
using Microsoft.AspNetCore.Mvc;
using Swashbuckle.AspNetCore.Annotations;
using Swashbuckle.AspNetCore.SwaggerGen;
using Newtonsoft.Json;
using System.ComponentModel.DataAnnotations;
using IO.Swagger.Attributes;

using Microsoft.AspNetCore.Authorization;

namespace IO.Swagger.Controllers
{
    /// <summary>
    /// 
    /// </summary>
    [ApiController]
    public class FaultHandlingApiController : ControllerBase
    {
        /// <summary>
        /// 
        /// </summary>
        /// <remarks>This method deletes all fault entries of an entity.</remarks>
        /// <param name="entityCollection">Path to an entity collection</param>
        /// <param name="entityId">ID of an entity</param>
        /// <param name="scope">The scope (e.g. user-defined error memories) for which faults are retrieved. The capability description defines which scopes are supported.</param>
        /// <response code="204">The request was successful.</response>
        /// <response code="0">An unexpected request occurred</response>
        [HttpDelete]
        [Route("/{entity-collection}/{entity-id}/faults")]
        [ValidateModelState]
        [SwaggerOperation("DeleteAllFaults")]
        public virtual IActionResult DeleteAllFaults([FromRoute][Required] Object entityCollection, [FromRoute][Required] Object entityId, [FromQuery] Object scope)
        {
            //TODO: Uncomment the next line to return response 204 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            return StatusCode(204);

            //TODO: Uncomment the next line to return response 0 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(0);


            throw new NotImplementedException();
        }

        /// <summary>
        /// 
        /// </summary>
        /// <remarks>This method deletes the given fault entry from an entity.</remarks>
        /// <param name="entityCollection">Path to an entity collection</param>
        /// <param name="entityId">ID of an entity</param>
        /// <param name="faultCode">The scope (e.g. user-defined error memories) for which faults are retrieved. The capability description defines which scopes are supported.</param>
        /// <response code="204">Fault entry has been deleted successfully.</response>
        /// <response code="404">The SOVD client tried to delete a fault entry which is not defined for the entity.</response>
        /// <response code="409">An unexpected request occurred</response>
        /// <response code="501">The request for deleting single fault entries is not supported by the entity.</response>
        /// <response code="0">An unexpected request occurred</response>
        [HttpDelete]
        [Route("/{entity-collection}/{entity-id}/faults/{fault-code}")]
        [ValidateModelState]
        [SwaggerOperation("DeleteFaultById")]
        public virtual IActionResult DeleteFaultById([FromRoute][Required] Object entityCollection, [FromRoute][Required] Object entityId, [FromRoute][Required] Object faultCode)
        {
            //TODO: Uncomment the next line to return response 204 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            return StatusCode(204);

            //TODO: Uncomment the next line to return response 404 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(404);

            //TODO: Uncomment the next line to return response 409 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(409);

            //TODO: Uncomment the next line to return response 501 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(501);

            //TODO: Uncomment the next line to return response 0 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(0);


            throw new NotImplementedException();
        }

        /// <summary>
        /// 
        /// </summary>
        /// <remarks>This method provides the details for a fault-code.</remarks>
        /// <param name="entityCollection">Path to an entity collection</param>
        /// <param name="entityId">ID of an entity</param>
        /// <param name="faultCode">The scope (e.g. user-defined error memories) for which faults are retrieved. The capability description defines which scopes are supported.</param>
        /// <param name="includeSchema">Specifies whether the response should include schema information or not. Default: false</param>
        /// <response code="200">Provide the details for a fault-code.</response>
        /// <response code="0">An unexpected request occurred</response>
        [HttpGet]
        [Route("/{entity-collection}/{entity-id}/faults/{fault-code}")]
        [ValidateModelState]
        [SwaggerOperation("GetFaultById")]
        public virtual IActionResult GetFaultById([FromRoute][Required] Object entityCollection, [FromRoute][Required] Object entityId, [FromRoute][Required] Object faultCode, [FromQuery] Object includeSchema)
        {
            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            return StatusCode(200);

            //TODO: Uncomment the next line to return response 0 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(0);


            throw new NotImplementedException();
        }

        /// <summary>
        /// 
        /// </summary>
        /// <remarks>This method provides the fault entries which are detected for an entity, if no query parameter is used. The returned fault entries may be filtered by their status.</remarks>
        /// <param name="entityCollection">Path to an entity collection</param>
        /// <param name="entityId">ID of an entity</param>
        /// <param name="includeSchema">Specifies whether the response should include schema information or not. Default: false</param>
        /// <param name="statusKey">Filters the available elements based on a key from the status. The provided value is a full match against the key and value of the attribute status in the response. For filtering by multiple status, the parameter is repeated (0 .. *). If multiple status are provided, they are logically OR combined. The keys supported by an entity are defined in the capability description of the entity.</param>
        /// <param name="severity">Filter fault entries by their severity. Any fault with a severity below the given number will be included in the response.</param>
        /// <param name="scope">The scope (e.g. user-defined error memories) for which faults are retrieved. The capability description defines which scopes are supported.</param>
        /// <response code="200">Retrieve list of all faults provided by the entity</response>
        /// <response code="0">An unexpected request occurred</response>
        [HttpGet]
        [Route("/{entity-collection}/{entity-id}/faults")]
        [ValidateModelState]
        [SwaggerOperation("GetFaults")]
        public virtual IActionResult GetFaults([FromRoute][Required] Object entityCollection, [FromRoute][Required] Object entityId, [FromQuery] Object includeSchema, [FromQuery] Object statusKey, [FromQuery] Object severity, [FromQuery] Object scope)
        {
            var jsonResponse = new
            {
                items = new[]
                {
                new {
                    code = "C11087",
                    scope = "",
                    display_code = "U0110-87",
                    fault_name = "Lost Communication With Drive Motor Control Module A",
                    fault_translation_id = "",
                    status = new {
                        aggregated_status = "Pending",
                        confirmedDTC = "0",
                        mask = "27",
                        pendingDTC = "1",
                        testFailed = "1",
                        testFailedSinceLastClear = "1",
                        testFailedThisOperationCycle = "1",
                        testNotCompletedSinceLastClear = "0",
                        testNotCompletedThisOperationCycle = "0",
                        warningIndicatorRequested = "0"
                    },
                    symptom = "missing message"
                },
                new {
                    code = "C06487",
                    scope = "",
                    display_code = "U0064-87",
                    fault_name = "Vehicle Communication Bus E",
                    fault_translation_id = "",
                    status = new {
                        aggregated_status = "Pending",
                        confirmedDTC = "0",
                        mask = "27",
                        pendingDTC = "1",
                        testFailed = "1",
                        testFailedSinceLastClear = "1",
                        testFailedThisOperationCycle = "1",
                        testNotCompletedSinceLastClear = "0",
                        testNotCompletedThisOperationCycle = "0",
                        warningIndicatorRequested = "0"
                    },
                    symptom = "missing message"
            },
                },
                Schema = ""
            };

            return Ok(jsonResponse); // ASP.NET Core will serialize to JSON automatically

            //TODO: Uncomment the next line to return response 0 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(0);


            throw new NotImplementedException();
        }
    }
}