# aspnetcore-server-generated/src/IO.Swagger/wwwroot/swagger-original.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "Definition of ASAM SOVD API",
    "version" : "1.0.0",
    "title" : "Service-Oriented Vehicle Diagnostic API",
    "contact" : {
      "url" : "https://www.asam.net/project-detail/sovd-service-oriented-vehicle-diagnostics/"
    }
  },
...
  "paths" : {
...
    "/{entity-collection}/{entity-id}/faults" : {
      "get" : {
        "tags" : [ "fault-handling" ],
        "description" : "This method provides the fault entries which are detected for an entity, if no query parameter is used. The returned fault entries may be filtered by their status.",
        "operationId" : "getFaults",
        "parameters" : [ {
          "name" : "entity-collection",
          "in" : "path",
          "description" : "Path to an entity collection",
          "required" : true
        }, {
          "name" : "entity-id",
          "in" : "path",
          "description" : "ID of an entity",
          "required" : true
        }, {
          "name" : "include-schema",
          "in" : "query",
          "description" : "Specifies whether the response should include schema information or not. Default: false",
          "required" : false
        }, {
          "name" : "status[key]",
          "in" : "query",
          "description" : "Filters the available elements based on a key from the status. The provided value is a full match against the key and value of the attribute status in the response. For filtering by multiple status, the parameter is repeated (0 .. *). If multiple status are provided, they are logically OR combined. The keys supported by an entity are defined in the capability description of the entity.",
          "required" : false
        }, {
          "name" : "severity",
          "in" : "query",
          "description" : "Filter fault entries by their severity. Any fault with a severity below the given number will be included in the response.",
          "required" : false
        }, {
          "name" : "scope",
          "in" : "query",
          "description" : "The scope (e.g. user-defined error memories) for which faults are retrieved. The capability description defines which scopes are supported.",
          "required" : false
        } ],
        "responses" : {
          "200" : {
            "description" : "Retrieve list of all faults provided by the entity"
          },
          "default" : {
            "description" : "An unexpected request occurred"
          }
        }
      },
      "delete" : {
        "tags" : [ "fault-handling" ],
        "description" : "This method deletes all fault entries of an entity.",
        "operationId" : "deleteAllFaults",
        "parameters" : [ {
          "name" : "entity-collection",
          "in" : "path",
          "description" : "Path to an entity collection",
          "required" : true
        }, {
          "name" : "entity-id",
          "in" : "path",
          "description" : "ID of an entity",
          "required" : true
        }, {
          "name" : "scope",
          "in" : "query",
          "description" : "The scope (e.g. user-defined error memories) for which faults are retrieved. The capability description defines which scopes are supported.",
          "required" : false
        } ],
        "responses" : {
          "204" : {
            "description" : "The request was successful."
          },
          "default" : {
            "description" : "An unexpected request occurred"
          }
        }
      }
    },
    "/{entity-collection}/{entity-id}/faults/{fault-code}" : {
      "get" : {
        "tags" : [ "fault-handling" ],
        "description" : "This method provides the details for a fault-code.",
        "operationId" : "getFaultById",
        "parameters" : [ {
          "name" : "entity-collection",
          "in" : "path",
          "description" : "Path to an entity collection",
          "required" : true
        }, {
          "name" : "entity-id",
          "in" : "path",
          "description" : "ID of an entity",
          "required" : true
        }, {
          "name" : "fault-code",
          "in" : "path",
          "description" : "The scope (e.g. user-defined error memories) for which faults are retrieved. The capability description defines which scopes are supported.",
          "required" : true
        }, {
          "name" : "include-schema",
          "in" : "query",
          "description" : "Specifies whether the response should include schema information or not. Default: false",
          "required" : false
        } ],
        "responses" : {
          "200" : {
            "description" : "Provide the details for a fault-code."
          },
          "default" : {
            "description" : "An unexpected request occurred"
          }
        }
      },
      "delete" : {
        "tags" : [ "fault-handling" ],
        "description" : "This method deletes the given fault entry from an entity.",
        "operationId" : "deleteFaultById",
        "parameters" : [ {
          "name" : "entity-collection",
          "in" : "path",
          "description" : "Path to an entity collection",
          "required" : true
        }, {
          "name" : "entity-id",
          "in" : "path",
          "description" : "ID of an entity",
          "required" : true
        }, {
          "name" : "fault-code",
          "in" : "path",
          "description" : "The scope (e.g. user-defined error memories) for which faults are retrieved. The capability description defines which scopes are supported.",
          "required" : true
        } ],
        "responses" : {
          "204" : {
            "description" : "Fault entry has been deleted successfully."
          },
          "404" : {
            "description" : "The SOVD client tried to delete a fault entry which is not defined for the entity."
          },
          "409" : {
            "description" : "An unexpected request occurred"
          },
          "501" : {
            "description" : "The request for deleting single fault entries is not supported by the entity."
          },
          "default" : {
            "description" : "An unexpected request occurred"
          }
        }
      }
    },
...
  },
...
}